pyproject.toml
[project]
name = "telgram-my-util-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.120.0",
    "onnxruntime>=1.23.2",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
    "rembg>=2.0.67",
    "yt-dlp>=2025.10.22",
]
